+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - **General** Use the same datetime for creating and reading the changelog files
 - **Package Handler** Fix wrong compare-url-tags-arch URL for the changelog file
 - **Package Handler** Fix wrong compare-url-tags-origin URL for the changelog file when facing for example KDE packages
+- **Package Handler** Correct docstring of find_intermediate_tags
 
 # 1.0.8 (2025-06-15)